+++ extensions/local-latex-ocr/src/__tests__/tokenizer.test.ts
@@ -7,11 +7,7 @@ import { LatexTokenizer } from "../lib/tokenizer";
 const temporaryDirectories: string[] = [];
 
 afterEach(async () => {
-  await Promise.all(
-    temporaryDirectories
-      .splice(0)
-      .map((directory) => rm(directory, { recursive: true, force: true })),
-  );
+  await Promise.all(temporaryDirectories.splice(0).map((directory) => rm(directory, { recursive: true, force: true })));
 });
 
 describe("LatexTokenizer", () => {